requirements.txt
monai==0.6.0
tqdm==4.59.0
einops==0.3.0
nibabel==3.1.1
pytorch-lightning==1.4.0
lightning-bolts==0.3.4
matplotlib
numpy
boto3
